requirements.txt
pyyaml
requests
nltk